python/eg_cloud_foundry_ssl_requests/requirements.txt
Flask-RESTful==0.3.5
requests==2.10.0

